% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_create_graphql_api}
\alias{appsync_create_graphql_api}
\title{Creates a GraphqlApi object}
\usage{
appsync_create_graphql_api(name, logConfig, authenticationType,
  userPoolConfig, openIDConnectConfig)
}
\arguments{
\item{name}{[required] A user-supplied name for the \code{GraphqlApi}.}

\item{logConfig}{The Amazon CloudWatch Logs configuration.}

\item{authenticationType}{[required] The authentication type: API key, AWS IAM, or Amazon Cognito user pools.}

\item{userPoolConfig}{The Amazon Cognito user pool configuration.}

\item{openIDConnectConfig}{The OpenID Connect configuration.}
}
\description{
Creates a \code{GraphqlApi} object.
}
\section{Request syntax}{
\preformatted{appsync$create_graphql_api(
  name = "string",
  logConfig = list(
    fieldLogLevel = "NONE"|"ERROR"|"ALL",
    cloudWatchLogsRoleArn = "string"
  ),
  authenticationType = "API_KEY"|"AWS_IAM"|"AMAZON_COGNITO_USER_POOLS"|"OPENID_CONNECT",
  userPoolConfig = list(
    userPoolId = "string",
    awsRegion = "string",
    defaultAction = "ALLOW"|"DENY",
    appIdClientRegex = "string"
  ),
  openIDConnectConfig = list(
    issuer = "string",
    clientId = "string",
    iatTTL = 123,
    authTTL = 123
  )
)
}
}

\keyword{internal}
