% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_delete_graphql_api}
\alias{appsync_delete_graphql_api}
\title{Deletes a GraphqlApi object}
\usage{
appsync_delete_graphql_api(apiId)
}
\arguments{
\item{apiId}{[required] The API ID.}
}
\description{
Deletes a \code{GraphqlApi} object.
}
\section{Request syntax}{
\preformatted{appsync$delete_graphql_api(
  apiId = "string"
)
}
}

\keyword{internal}
