% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_list_functions}
\alias{appsync_list_functions}
\title{List multiple functions}
\usage{
appsync_list_functions(apiId, nextToken, maxResults)
}
\arguments{
\item{apiId}{[required] The GraphQL API ID.}

\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}

\item{maxResults}{The maximum number of results you want the request to return.}
}
\description{
List multiple functions.
}
\section{Request syntax}{
\preformatted{appsync$list_functions(
  apiId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
