% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appsync_operations.R
\name{appsync_list_graphql_apis}
\alias{appsync_list_graphql_apis}
\title{Lists your GraphQL APIs}
\usage{
appsync_list_graphql_apis(nextToken, maxResults)
}
\arguments{
\item{nextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}

\item{maxResults}{The maximum number of results you want the request to return.}
}
\description{
Lists your GraphQL APIs.
}
\section{Request syntax}{
\preformatted{appsync$list_graphql_apis(
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
